Posix build (#1664)
* Changes to support Posix build target

The Posix build can serve as a base for automated tests.

The vast majority of the changes were extant problems caught
by a pickier compiler, including

* Use snprintf instead of sprintf for safety
* Missing or spurious "override" on virtual methods
* A couple of buffer safety issues
* Nonportable use of variable length arrays
* Case-correct spelling of freertos Task.H and Queue.h
* Type mismatches in printf arguments

I had to change the way that undefinedPin and errorPin were defined
to ensure that they are initialized prior to being used, which
showed up with a different compiler.

The changes that were directly related to Posix were quite
small.  A WinConsole.cpp and ExceptionHelper.cpp were moved
from capture/ to (new) windows-x86/, and posix/Console.cpp
was created to handle keyboard input.

* Spurious "private:"

* Task.h and Queue.h should be lowercase

* Still fixing capitalization

* This time for sure

* Reversed preliminary commit of Channel.cpp fix

* Put errorPin and undefinedPin in the Pins namespace

+++ FluidNC/capture/fnc_uart.cpp
@@ -29,7 +29,7 @@ int uart_buflen(uint32_t uart_num) {
 
 extern int inchar();
 
-int uart_read(uint32_t uart_num, uint8_t* buf, size_t len, uint32_t timeout_ms) {
+int uart_read(uint32_t uart_num, uint8_t* buf, uint32_t len, uint32_t timeout_ms) {
     auto        key = uart_key(uart_num);
     const auto& val = Inputs::instance().get(key);
     auto        max = std::min(size_t(len), val.size());
